+++ src/Exception/ExceptionMessages.php
@@ -38,7 +38,7 @@ class ExceptionMessages
         10114 => 'Attempting to set a read-only property',
         10115 => 'Unexpected class encountered while preparing row',
         10116 => 'Attempting to get an undefined property',
-        10117 => 'ModelCollection is referencing ad undefined, non-primitive class'
+        10117 => 'ModelCollection is referencing ad undefined, non-primitive class',
     ];
 
     /**
