+++ tests/SageOne/Models/ExampleModelTest.php
@@ -489,4 +489,96 @@ class ExampleModelTest extends BaseModelTest
             $result
         );
     }
+
+    public function testCollectionWithoutClass()
+    {
+        $this->expectException(ModelException::class);
+        $this->expectExceptionMessage('Model "Example" class "DarrynTen\SageOne\Models\InvalidClass" ModelCollection is referencing ad undefined, non-primitive class');
+        $this->expectExceptionCode(10117);
+
+        $exampleModel = new Example($this->config);
+        $exampleBadFields = [
+            'someCollection' => [
+                'type' => 'ModelCollection',
+                'nullable' => false,
+                'readonly' => false,
+                'class' => 'InvalidClass'
+            ],
+        ];
+
+        $reflection = new ReflectionClass($exampleModel);
+        $reflectedModel = $reflection->getProperty('fields');
+        $reflectedModel->setAccessible(true);
+        $reflectedModel->setValue($exampleModel, $exampleBadFields);
+
+        $data = json_decode(file_get_contents(__DIR__ . '/../../mocks/Example/GET_Example_Get_xx.json'));
+        $exampleModel->loadResult($data);
+    }
+
+    public function testAttributes()
+    {
+        $this->verifyAttributes(Example::class, [
+            'id' => [
+                'type' => 'integer',
+                'nullable' => false,
+                'readonly' => false,
+            ],
+            'exampleWithCamel' => [
+                'type' => 'string',
+                'nullable' => true,
+                'readonly' => true,
+            ],
+            'stringRange' => [
+                'type' => 'string',
+                'nullable' => true,
+                'readonly' => false,
+                'min' => 2,
+                'max' => 10,
+            ],
+            'stringWithDefault' => [
+                'type' => 'string',
+                'nullable' => true,
+                'readonly' => false,
+                'default' => 'some default value',
+            ],
+            'stringWithNullDefault' => [
+                'type' => 'string',
+                'nullable' => true,
+                'readonly' => false,
+                'default' => null
+            ],
+            'integerRange' => [
+                'type' => 'integer',
+                'nullable' => false,
+                'readonly' => false,
+                'min' => 1,
+                'max' => 2147483647,
+            ],
+            'someBoolean' => [
+                'type' => 'boolean',
+                'nullable' => false,
+                'readonly' => false,
+            ],
+            'requiredString' => [
+                'type' => 'string',
+                'nullable' => false,
+                'readonly' => false,
+                'required' => true,
+            ],
+            'emailAddress' => [
+                'type' => 'string',
+                'nullable' => false,
+                'readonly' => false,
+                'min' => 0,
+                'max' => 100,
+                'regex' => "/^[A-Za-z0-9,!#\$%&'\*\+\/=\?\^_`\{\|}~-]+(\.[A-Za-z0-9,!#\$%&'\*\+\/=\?\^_`\{\|}~-]+)*@[A-Za-z0-9-]+(\.[A-Za-z0-9-]+)*\.([A-Za-z]{2,})$/"
+            ],
+            'someCollection' => [
+                'type' => 'ModelCollection',
+                'class' => 'ExampleCategory',
+                'nullable' => false,
+                'readonly' => false
+            ]
+        ]);
+    }
 }
